requirements.txt
Django==2.2.7
django-admin==2.0.0
django-appconf==1.0.3
django-compressor==2.3
django-excel-response2==2.0.8
django-libsass==0.7
django-six==1.0.4
libsass==0.19.4
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
screen==1.0.1
six==1.13.0
sqlparse==0.3.0
xlwt==1.3.0
django-ckeditor==5.9.0
Pillow==7.1.2